fixes and put models in diff folder
